fixed; 5a6202a44daf1e9bd33379581d56abbfccfa0f02

+++ services/horizon/internal/config.go
@@ -43,7 +43,7 @@ type Config struct {
 	SkipCursorUpdate bool
 
 	// BaseFee sets the base fee of network
-	BaseFee int
+	BaseFee uint64
 
 	// InflationAccount only can occur the inflation operation
 	InflationAccount xdr.AccountId
